Implemented some basic JSoup
Obviously, the scope of this project has changed. Rather than create the
application from the ground up, it is simpler and more productive to
create a streamlined form the the online version. I've moved to JSoup
(keeping the old code for awhile) for processing the pages. Noted bugs:
- The alert buttons are not removed using conventional JSoup functions.
Probably added through javascript, and will have to block that somehow
- The late stages of the hexagon puzzles don't render properly. This is
on my phone (1920x1080), and could have a number of causes (CSS
attributes, webview settings, etc)

At this point of the project, my main goals are to:
- Finish sanitizing puzzle pages
- Add activity pages for puzzle selection
- Finalization + installation (I've never considered putting something
on the Google Play store, and this may or may not materialize)